# mak_it_tz/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Admin panel",
        "description": "Admin panel for management",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Заказы"
                ],
                "summary": "Get a list of orders",
                "operationId": "order.index",
                "responses": {
                    "200": {
                        "description": "List of orders"
                    },
                    "401": {
                        "description": "Unauthorized - Missing or invalid token"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Заказы"
                ],
                "summary": "Create a new order",
                "operationId": "order.store",
                "requestBody": {
                    "description": "Order data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "product_id",
                                    "quantity",
                                    "total_price"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "ID of the user placing the order",
                                        "type": "integer"
                                    },
                                    "product_id": {
                                        "description": "ID of the product being ordered",
                                        "type": "integer"
                                    },
                                    "quantity": {
                                        "description": "Quantity of the product",
                                        "type": "integer"
                                    },
                                    "total_price": {
                                        "description": "Total price of the order",
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully"
                    },
                    "400": {
                        "description": "Bad request - Invalid input data"
                    },
                    "401": {
                        "description": "Unauthorized - Missing or invalid token"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Заказы"
                ],
                "summary": "Get information about an order",
                "operationId": "order.show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the order",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order information"
                    },
                    "401": {
                        "description": "Unauthorized - Missing or invalid token"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Заказы"
                ],
                "summary": "Update an existing order",
                "operationId": "order.update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the order to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated order data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "product_id",
                                    "quantity",
                                    "total_price"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "ID of the user placing the order",
                                        "type": "integer"
                                    },
                                    "product_id": {
                                        "description": "ID of the product being ordered",
                                        "type": "integer"
                                    },
                                    "quantity": {
                                        "description": "Quantity of the product",
                                        "type": "integer"
                                    },
                                    "total_price": {
                                        "description": "Total price of the order",
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order updated successfully"
                    },
                    "400": {
                        "description": "Bad request - Invalid input data"
                    },
                    "401": {
                        "description": "Unauthorized - Missing or invalid token"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Заказы"
                ],
                "summary": "Delete an existing order",
                "operationId": "order.destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the order to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Order successfully deleted"
                    },
                    "401": {
                        "description": "Unauthorized - Missing or invalid token"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/product/{id}": {
            "get": {
                "tags": [
                    "Продукты"
                ],
                "summary": "Получить информацию о продукте",
                "operationId": "product.show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID продукта",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о продукте"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Продукты"
                ],
                "summary": "Удалить продукт",
                "operationId": "product.destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID продукта",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Продукт успешно удален"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}